+++ pytorch_lightning/loops/optimization/optimizer_loop.py
@@ -123,8 +123,11 @@ class OptimizerLoop(Loop):
 
         closure = self._make_closure(split_batch, batch_idx, opt_idx, optimizer)
 
-        # when the training type plugin handles accumulation, we want to always call the optimizer step
-        if not self.trainer.accelerator.handles_accumulate_grad_batches and self.trainer.fit_loop.should_accumulate():
+        if (
+            # when the training type plugin handles accumulation, we want to always call the optimizer step
+            not self.trainer.training_type_plugin.handles_accumulate_grad_batches
+            and self.trainer.fit_loop.should_accumulate()
+        ):
             # For gradient accumulation
 
             # -------------------